Split input/output out of main.go
